+++ sei-ibc-go/modules/core/02-client/client/cli/cli.go
@@ -20,6 +20,7 @@ func GetQueryCmd() *cobra.Command {
 	queryCmd.AddCommand(
 		GetCmdQueryClientStates(),
 		GetCmdQueryClientState(),
+		GetCmdQueryClientStatus(),
 		GetCmdQueryConsensusStates(),
 		GetCmdQueryConsensusState(),
 		GetCmdQueryHeader(),